fix: keep the shared shallow-equal react-free and the transport pair aligned
