+++ examples/with-assistant-transport/app/MyRuntimeProvider.tsx
@@ -64,6 +64,7 @@ export function MyRuntimeProvider({ children }: MyRuntimeProviderProps) {
     initialState: {
       messages: [],
     },
+    protocol: "assistant-transport",
     api: process.env.NEXT_PUBLIC_API_URL || "http://localhost:8010/assistant",
     converter,
     headers: async () => ({
