+++ packages/store/src/utils/useShallowStable.ts
@@ -1,25 +1,7 @@
 import { useMemo, useRef } from "react";
+import { shallowEqual } from "./shallow-equal";
 
-export const shallowEqual = (a: object, b: object): boolean => {
-  if (Array.isArray(a) !== Array.isArray(b)) return false;
-  if (Array.isArray(a) && Array.isArray(b)) {
-    return (
-      a.length === b.length && a.every((value, i) => Object.is(value, b[i]))
-    );
-  }
-  const aKeys = Object.keys(a);
-  return (
-    aKeys.length === Object.keys(b).length &&
-    aKeys.every(
-      (key) =>
-        Object.hasOwn(b, key) &&
-        Object.is(
-          (a as Record<string, unknown>)[key],
-          (b as Record<string, unknown>)[key],
-        ),
-    )
-  );
-};
+export { shallowEqual };
 
 export const useShallowStable = <T extends object>(value: T): T => {
   const cell = useMemo(() => ({}) as { v?: T }, []);
